+++ msalign/__init__.py
@@ -17,7 +17,7 @@ def msalign(
     array: np.ndarray,
     peaks: List,
     method: str = "cubic",
-    width: int = 10,
+    width: float = 10,
     ratio: float = 2.5,
     resolution: int = 100,
     iterations: int = 5,
@@ -27,7 +27,6 @@ def msalign(
     return_shifts: bool = False,
     align_by_index: bool = False,
     only_shift: bool = False,
-    quick_shift: bool = False,
 ):
     aligner = Aligner(
         x,
@@ -44,10 +43,9 @@ def msalign(
         return_shifts=return_shifts,
         align_by_index=align_by_index,
         only_shift=only_shift,
-        quick_shift=quick_shift,
     )
     aligner.run()
-    return aligner.align()
+    return aligner.apply()
 
 
 msalign.__doc__ = Aligner.__doc__