+++ msalign/align.py
@@ -1,12 +1,12 @@
 """Main alignment class"""
 import logging
 import time
+import typing as ty
 import warnings
-from typing import List
 
 import numpy as np
 
-from .utilities import check_xy, convert_peak_values_to_index, format_time, generate_function, shift, time_loop
+from .utilities import check_xy, convert_peak_values_to_index, generate_function, shift, time_loop
 
 METHODS = ["pchip", "zero", "slinear", "quadratic", "cubic", "linear"]
 LOGGER = logging.getLogger(__name__)
@@ -15,23 +15,26 @@ LOGGER = logging.getLogger(__name__)
 class Aligner:
     """Main alignment class"""
 
+    _method, _gaussian_ratio, _gaussian_resolution, _gaussian_width, _n_iterations = None, None, None, None, None
+    _corr_sig_l, _corr_sig_x, _corr_sig_y, _reduce_range_factor, _scale_range = None, None, None, None, None
+    _search_space, _computed = None, False
+
     def __init__(
         self,
         x: np.ndarray,
-        array: np.ndarray,
-        peaks: List,
+        array: ty.Optional[np.ndarray],
+        peaks: ty.Iterable[float],
         method: str = "cubic",
-        width: int = 10,
+        width: float = 10,
         ratio: float = 2.5,
         resolution: int = 100,
         iterations: int = 5,
         grid_steps: int = 20,
-        shift_range: List = None,
-        weights: List = None,
+        shift_range: ty.Optional[ty.Tuple[int, int]] = None,
+        weights: ty.Optional[ty.List[float]] = None,
         return_shifts: bool = False,
         align_by_index: bool = False,
         only_shift: bool = False,
-        quick_shift: bool = False,
     ):
         """Signal calibration and alignment by reference peaks
 
@@ -85,48 +88,47 @@ class Aligner:
             decide whether shift parameter `shift_opt` should also be returned. Default: False
         align_by_index : bool
             decide whether alignment should be done based on index rather than `xvals` array. Default: False
-
-        Returns
-        -------
-        zvals_out : numpy array
-            calibrated array
-        shift_opt : numpy array (optional)
-            amount of shift for each signal. Only returned if return_shift is set to True
         """
-        t_start = time.time()
-        # set input
         self.x = np.asarray(x)
-        self.array = check_xy(self.x, np.asarray(array))
+        if array is not None:
+            self.array = check_xy(self.x, np.asarray(array))
+        else:
+            self.array = np.empty((0, len(self.x)))
+
+        self.n_signals = self.array.shape[0]
+        self.array_aligned = np.zeros_like(self.array)
         self.peaks = list(peaks)
 
         # set attributes
-        self.n_signals = self.array.shape[0]
         self.n_peaks = len(self.peaks)
 
         # accessible attributes
-        self.array_aligned = np.zeros_like(self.array)
         self.scale_opt = np.ones((self.n_signals, 1), dtype=np.float32)
         self.shift_opt = np.zeros((self.n_signals, 1), dtype=np.float32)
         self.shift_values = np.zeros_like(self.shift_opt)
 
-        # interpolation method
-        self._method = method
-        # std dev of the Gaussian pulses (in X)
-        self._gaussian_width = width
-        # sets the width of the windows use at every pulse
-        self._gaussian_ratio = ratio
-        # resolution of every Gaussian pulse (number of points)
-        self._gaussian_resolution = resolution
-        # increase to improve accuracy
-        self._n_iterations = iterations
-        # size of grid for exhaustive search
-        self._grid_steps = grid_steps
-        # initial shift range
+        self.method = method
+        self.gaussian_ratio = ratio
+        self.gaussian_resolution = resolution
+        self.gaussian_width = width
+        self.n_iterations = iterations
+        self.grid_steps = grid_steps
         if shift_range is None:
             shift_range = [-100, 100]
-        self._shift_range = np.asarray(shift_range)
+        self.shift_range = shift_range
+        if weights is None:
+            weights = np.ones(self.n_peaks)
+        self.weights = weights
+
         # return shift vector
         self._return_shifts = return_shifts
+        # If the number of points is equal to 1, then only shift
+        if self.n_peaks == 1:
+            only_shift = True
+        if only_shift and not align_by_index:
+            align_by_index = True
+            LOGGER.warning("Only computing shifts - changed `align_by_index` to `True`.")
+
         # align signals by index rather than peak value
         self._align_by_index = align_by_index
         # align by index - rather than aligning to arbitrary non-integer values in the xvals, you can instead
@@ -136,189 +138,221 @@ class Aligner:
             self.x = np.arange(self.x.shape[0])
             LOGGER.debug(f"Aligning by index - peak positions: {self.peaks}")
         self._only_shift = only_shift
-        if self.n_peaks == 1:
-            self._only_shift = True
-        # weights
-        if weights is None:
-            weights = np.ones(self.n_peaks)
-        self._weights = weights
 
-        # enable quick realignment
-        self._quick_shift = quick_shift
+        self._initialize()
 
-        # private attributes
-        self._corr_sig_l = None
-        self._corr_sig_x = None
-        self._corr_sig_y = None
-        self._reduce_range_factor = None
-        self._scale_range = None
-        self._search_space = None
-        self._computed = False
+    @property
+    def method(self):
+        """Interpolation method."""
+        return self._method
 
-        # validate inputs
-        self.validate()
+    @method.setter
+    def method(self, value: str):
+        if value not in METHODS:
+            raise ValueError(f"Method `{value}` not found in the method options: {METHODS}")
+        self._method = value
 
-        # prepare
-        self.prepare()
-        LOGGER.debug(f"Initialized in {format_time(time.time()-t_start)}")
+    @property
+    def gaussian_ratio(self):
+        """Gaussian ratio."""
+        return self._gaussian_ratio
 
-    def validate(self):
-        """Ensures the user-set parameters are correct"""
-        # check user-specified parameters
-        if self._method not in METHODS:
-            raise ValueError(f"Method `{self._method}` not found in the method options: {METHODS}")
-        if not isinstance(self._weights, (list, set, np.ndarray)) or len(self._weights) != self.n_peaks:
-            raise ValueError("Number of weights does not match number of peaks")
-        if len(self._shift_range) != 2:
+    @gaussian_ratio.setter
+    def gaussian_ratio(self, value: float):
+        if value <= 0:
+            raise ValueError("Value of 'ratio' must be above 0!")
+        self._gaussian_ratio = value
+
+    @property
+    def gaussian_resolution(self):
+        """Gaussian resolution of every Gaussian pulse (number of points)."""
+        return self._gaussian_resolution
+
+    @gaussian_resolution.setter
+    def gaussian_resolution(self, value: float):
+        if value <= 0:
+            raise ValueError("Value of 'resolution' must be above 0!")
+        self._gaussian_resolution = value
+
+    @property
+    def gaussian_width(self):
+        """Width of the Gaussian pulse in std dev of the Gaussian pulses (in X)."""
+        return self._gaussian_width
+
+    @gaussian_width.setter
+    def gaussian_width(self, value: float):
+        self._gaussian_width = value
+
+    @property
+    def n_iterations(self):
+        """Total number of iterations - increase to improve accuracy."""
+        return self._n_iterations
+
+    @n_iterations.setter
+    def n_iterations(self, value: int):
+        if value < 1 or not isinstance(value, int):
+            raise ValueError("Value of 'iterations' must be above 0 and be an integer!")
+        self._n_iterations = value
+
+    @property
+    def grid_steps(self):
+        """Total number of iterations - increase to improve accuracy."""
+        return self._grid_steps
+
+    @grid_steps.setter
+    def grid_steps(self, value: int):
+        if value < 1 or not isinstance(value, int):
+            raise ValueError("Value of 'iterations' must be above 0 and be an integer!")
+        self._grid_steps = value
+
+    @property
+    def shift_range(self):
+        """Total number of iterations - increase to improve accuracy."""
+        return self._shift_range
+
+    @shift_range.setter
+    def shift_range(self, value: ty.Tuple[float, float]):
+        if len(value) != 2:
             raise ValueError(
                 "Number of 'shift_values' is not correct. Shift range accepts" " numpy array with two values."
             )
-        if np.diff(self._shift_range) == 0:
+        if np.diff(value) == 0:
             raise ValueError("Values of 'shift_values' must not be the same!")
-        if self._gaussian_ratio <= 0:
-            raise ValueError("Value of 'ratio' must be above 0!")
-        if self._n_iterations <= 0 or isinstance(self._n_iterations, float):
-            raise ValueError("Value of 'iterations' must be above 0 and be an integer!")
-        if self._grid_steps <= 0:
-            raise ValueError("Value of 'grid_steps' must be above 0!")
-        if self._gaussian_resolution <= 0:
-            raise ValueError("Value of 'resolution' must be above 0!")
-        if not isinstance(self._only_shift, bool):
-            raise ValueError("Value of 'only_shift' must be a boolean")
-        if not isinstance(self._return_shifts, bool):
-            raise ValueError("Value of 'return_shift' must be a boolean")
-        if not isinstance(self._align_by_index, bool):
-            raise ValueError("Value of 'align_by_index' must be a boolean")
-        if self._quick_shift and not self._only_shift:
-            raise ValueError(
-                "Cannot set `quick_shift` with rescaling since quick shift is accomplished by simply"
-                " moving spectra along the axis without interpolation"
-            )
-        if self._quick_shift and not self._align_by_index:
-            raise ValueError(
-                "Cannot set `quick_shift` without also setting the `align_by_index` to `True`. Quick shift"
-                " simply moves the spectra without interpolation so it relies on moving values by their"
-                " index which is not calculated correctly when using `true` x-axis values."
-            )
+        self._shift_range = np.asarray(value)
 
-    def _set_attr(self, attr: str, new_value, current_value):
-        """Set attribute on the alignment class while ignoring the request if the value is ``None``"""
-        if new_value is None or new_value == current_value:
-            return
-        setattr(self, attr, new_value)
+    @property
+    def weights(self):
+        """Total number of iterations - increase to improve accuracy."""
+        return self._weights
 
-    def prepare(self):
+    @weights.setter
+    def weights(self, value: ty.Optional[ty.Iterable[float]]):
+        if value is None:
+            value = np.ones(self.n_peaks)
+        if not isinstance(value, ty.Iterable):
+            raise ValueError("Weights must be provided as an iterable.")
+        if len(value) != self.n_peaks:
+            raise ValueError("Number of weights does not match the number of peaks.")
+        self._weights = np.asarray(value)
+
+    def _initialize(self):
         """Prepare dataset for alignment"""
-        t_start = time.time()
         # check that values for gaussian_width are valid
         gaussian_widths = np.zeros((self.n_peaks, 1))
         for i in range(self.n_peaks):
-            gaussian_widths[i] = self._gaussian_width
+            gaussian_widths[i] = self.gaussian_width
 
         # set the synthetic target signal
-        corr_sig_x = np.zeros((self._gaussian_resolution + 1, self.n_peaks))
-        corr_sig_y = np.zeros((self._gaussian_resolution + 1, self.n_peaks))
+        corr_sig_x = np.zeros((self.gaussian_resolution + 1, self.n_peaks))
+        corr_sig_y = np.zeros((self.gaussian_resolution + 1, self.n_peaks))
 
-        gaussian_resolution_range = np.arange(0, self._gaussian_resolution + 1)
+        gaussian_resolution_range = np.arange(0, self.gaussian_resolution + 1)
         for i in range(self.n_peaks):
-            left_l = self.peaks[i] - self._gaussian_ratio * gaussian_widths[i]  # noqa
-            right_l = self.peaks[i] + self._gaussian_ratio * gaussian_widths[i]  # noqa
-            corr_sig_x[:, i] = left_l + (gaussian_resolution_range * (right_l - left_l) / self._gaussian_resolution)
-            corr_sig_y[:, i] = self._weights[i] * np.exp(
+            left_l = self.peaks[i] - self.gaussian_ratio * gaussian_widths[i]  # noqa
+            right_l = self.peaks[i] + self.gaussian_ratio * gaussian_widths[i]  # noqa
+            corr_sig_x[:, i] = left_l + (gaussian_resolution_range * (right_l - left_l) / self.gaussian_resolution)
+            corr_sig_y[:, i] = self.weights[i] * np.exp(
                 -np.square((corr_sig_x[:, i] - self.peaks[i]) / gaussian_widths[i])  # noqa
             )
 
-        self._corr_sig_l = (self._gaussian_resolution + 1) * self.n_peaks
+        self._corr_sig_l = (self.gaussian_resolution + 1) * self.n_peaks
         self._corr_sig_x = corr_sig_x.flatten("F")
         self._corr_sig_y = corr_sig_y.flatten("F")
 
         # set reduce_range_factor to take 5 points of the previous ranges or half of
         # the previous range if grid_steps < 10
-        self._reduce_range_factor = min(0.5, 5 / self._grid_steps)
+        self._reduce_range_factor = min(0.5, 5 / self.grid_steps)
 
         # set scl such that the maximum peak can shift no more than the limits imposed by shift when scaling
-        self._scale_range = 1 + self._shift_range / max(self.peaks)
+        self._scale_range = 1 + self.shift_range / max(self.peaks)
 
         if self._only_shift:
             self._scale_range = np.array([1, 1])
 
-        # create the meshgrid only once
+        # create the mesh-grid only once
         mesh_a, mesh_b = np.meshgrid(
-            np.divide(np.arange(0, self._grid_steps), self._grid_steps - 1),
-            np.divide(np.arange(0, self._grid_steps), self._grid_steps - 1),
+            np.divide(np.arange(0, self.grid_steps), self.grid_steps - 1),
+            np.divide(np.arange(0, self.grid_steps), self.grid_steps - 1),
         )
         self._search_space = np.tile(
             np.vstack([mesh_a.flatten(order="F"), mesh_b.flatten(order="F")]).T, [1, self._n_iterations]
         )
-        LOGGER.debug(f"Prepared in {format_time(time.time() - t_start)}")
 
     def run(self, n_iterations: int = None):
         """Execute the alignment procedure for each signal in the 2D array and collate the shift/scale vectors"""
-        self._set_attr("_n_iterations", n_iterations, self._n_iterations)
+        self.n_iterations = n_iterations or self.n_iterations
         # iterate for every signal
         t_start = time.time()
 
-        _scale_range = np.array([-0.5, 0.5])
         # main loop: searches for the optimum values of Scale and Shift factors by search over a multi-resolution
         # grid, getting better at each iteration. Increasing the number of iterations improves the shift and scale
         # parameters
         for n_signal, y in enumerate(self.array):
-            # set to back to the user input arguments (or default)
-            _shift = self._shift_range
-            _scale = self._scale_range
-
-            # generate interpolation function for each signal - instantiation of the interpolator can be quite slow,
-            # so you can slightly increase the number of iterations without significant slowdown of the process
-            fcn = generate_function(self._method, self.x, y)
-
-            # iterate to estimate the shift and scale - at each iteration, the grid search is readjusted and the
-            # shift/scale values are optimized further
-            for n_iter in range(self._n_iterations):
-                # scale and shift search space
-                scale_grid = _scale[0] + self._search_space[:, (n_iter * 2) - 2] * np.diff(_scale)
-                shift_grid = _shift[0] + self._search_space[:, (n_iter * 2) + 1] * np.diff(_shift)
-                temp = (
-                    np.reshape(scale_grid, (scale_grid.shape[0], 1))
-                    * np.reshape(self._corr_sig_x, (1, self._corr_sig_l))
-                    + np.tile(shift_grid, [self._corr_sig_l, 1]).T
-                )
-                # interpolate at each iteration. Need to remove NaNs which can be introduced by certain (e.g.
-                # PCHIP) interpolator
-                temp = np.nan_to_num(fcn(temp.flatten("C")).reshape(temp.shape))
-
-                # determine the best position
-                i_max = np.dot(temp, self._corr_sig_y).argmax()
-
-                # save optimum value
-                self.scale_opt[n_signal] = scale_grid[i_max]
-                self.shift_opt[n_signal] = shift_grid[i_max]
-
-                # readjust grid for next iteration_reduce_range_factor
-                _scale = self.scale_opt[n_signal] + _scale_range * np.diff(_scale) * self._reduce_range_factor
-                _shift = self.shift_opt[n_signal] + _scale_range * np.diff(_shift) * self._reduce_range_factor
+            self.shift_opt[n_signal], self.scale_opt[n_signal] = self.compute(y)
         LOGGER.debug(f"Processed {self.n_signals} signals " + time_loop(t_start, self.n_signals + 1, self.n_signals))
         self._computed = True
 
-    def align(self, quick_shift: bool = None, return_shifts: bool = None):
+    def compute(self, y: np.ndarray) -> ty.Tuple[float, float]:
+        """Compute correction factors.
+
+        This function does not set value in any of the class attributes so can be used in a iterator where values
+        are computed lazily.
+        """
+        _scale_range = np.array([-0.5, 0.5])
+        scale_opt, shift_opt = 0.0, 1.0
+
+        # set to back to the user input arguments (or default)
+        _shift = self.shift_range.copy()
+        _scale = self._scale_range.copy()
+
+        # generate interpolation function for each signal - instantiation of the interpolator can be quite slow,
+        # so you can slightly increase the number of iterations without significant slowdown of the process
+        func = generate_function(self.method, self.x, y)
+
+        # iterate to estimate the shift and scale - at each iteration, the grid search is readjusted and the
+        # shift/scale values are optimized further
+        for n_iter in range(self.n_iterations):
+            # scale and shift search space
+            scale_grid = _scale[0] + self._search_space[:, (n_iter * 2) - 2] * np.diff(_scale)
+            shift_grid = _shift[0] + self._search_space[:, (n_iter * 2) + 1] * np.diff(_shift)
+            temp = (
+                np.reshape(scale_grid, (scale_grid.shape[0], 1)) * np.reshape(self._corr_sig_x, (1, self._corr_sig_l))
+                + np.tile(shift_grid, [self._corr_sig_l, 1]).T
+            )
+            # interpolate at each iteration. Need to remove NaNs which can be introduced by certain (e.g.
+            # PCHIP) interpolator
+            temp = np.nan_to_num(func(temp.flatten("C")).reshape(temp.shape))
+
+            # determine the best position
+            i_max = np.dot(temp, self._corr_sig_y).argmax()
+
+            # save optimum value
+            scale_opt = scale_grid[i_max]
+            shift_opt = shift_grid[i_max]
+
+            # readjust grid for next iteration_reduce_range_factor
+            _scale = scale_opt + _scale_range * np.diff(_scale) * self._reduce_range_factor
+            _shift = shift_opt + _scale_range * np.diff(_shift) * self._reduce_range_factor
+        return shift_opt, scale_opt
+
+    def apply(self, return_shifts: bool = None):
         """Align the signals against the computed values"""
         if not self._computed:
             warnings.warn("Aligning data without computing optimal alignment parameters", UserWarning)
-        self._set_attr("_quick_shift", quick_shift, self._quick_shift)
-        self._set_attr("_return_shifts", return_shifts, self._return_shifts)
+        self._return_shifts = return_shifts if return_shifts is not None else self._return_shifts
 
-        if self._quick_shift:
+        if self._only_shift:
             self.shift()
         else:
-            self.realign()
+            self.align()
 
         # return aligned data and shifts
         if self._return_shifts:
             return self.array_aligned, self.shift_values
-
         # only return data
         return self.array_aligned
 
-    def realign(self, shift_opt=None, scale_opt=None):
+    def align(self, shift_opt=None, scale_opt=None):
         """Realign array based on the optimized shift and scale parameters
 
         Parameters
@@ -337,14 +371,20 @@ class Aligner:
         # realign based on provided values
         for iteration, y in enumerate(self.array):
             # interpolate back to the original domain
-            fcn = generate_function(self._method, (self.x - shift_opt[iteration]) / scale_opt[iteration], y)
-            self.array_aligned[iteration] = np.nan_to_num(fcn(self.x))
+            self.array_aligned[iteration] = self._apply(y, shift_opt[iteration], scale_opt[iteration])
         self.shift_values = self.shift_opt
 
         LOGGER.debug(f"Re-aligned {self.n_signals} signals " + time_loop(t_start, self.n_signals + 1, self.n_signals))
 
+    def _apply(self, y: np.ndarray, shift_value: float, scale_value: float):
+        """Apply alignment correction to array `y`."""
+        func = generate_function(self.method, (self.x - shift_value) / scale_value, y)
+        return np.nan_to_num(func(self.x))
+
     def shift(self, shift_opt=None):
-        """Quickly shift array based on the optimized shift parameters
+        """Quickly shift array based on the optimized shift parameters.
+
+        This method does not interpolate but rather moves the data left and right without applying any scaling.
 
         Parameters
         ----------
@@ -357,7 +397,12 @@ class Aligner:
 
         # quickly shift based on provided values
         for iteration, y in enumerate(self.array):
-            self.array_aligned[iteration] = shift(y, -int(shift_opt[iteration]))
+            self.array_aligned[iteration] = self._shift(y, shift_opt[iteration])
         self.shift_values = shift_opt
 
         LOGGER.debug(f"Re-aligned {self.n_signals} signals " + time_loop(t_start, self.n_signals + 1, self.n_signals))
+
+    @staticmethod
+    def _shift(y: np.ndarray, shift_value: float):
+        """Apply shift correction to array `y`."""
+        return shift(y, -int(shift_value))